refactor tests to consolidate a single vectorize folder
