+++ tests/idiomatic/conftest.py
@@ -24,18 +24,8 @@ from astrapy.constants import VectorMetric
 
 TEST_COLLECTION_INSTANCE_NAME = "test_coll_instance"
 TEST_COLLECTION_NAME = "id_test_collection"
-TEST_SERVICE_COLLECTION_NAME = "test_service_collection"
 
 ASTRA_DB_SECONDARY_KEYSPACE = os.environ.get("ASTRA_DB_SECONDARY_KEYSPACE")
-
-
-def is_vector_service_available() -> bool:
-    return all(
-        [
-            "us-west-2" in os.environ.get("ASTRA_DB_API_ENDPOINT", ""),
-            "astra-dev.datastax.com" in os.environ.get("ASTRA_DB_API_ENDPOINT", ""),
-        ]
-    )
 
 
 @pytest.fixture(scope="session")
@@ -89,36 +79,11 @@ def sync_collection(
     sync_database.drop_collection(TEST_COLLECTION_NAME)
 
 
-@pytest.fixture(scope="session")
-def sync_service_collection(
-    astra_db_credentials_kwargs: AstraDBCredentials,
-    sync_database: Database,
-) -> Iterable[Collection]:
-    """An actual collection on DB, in the main namespace"""
-    collection = sync_database.create_collection(
-        TEST_SERVICE_COLLECTION_NAME,
-        metric=VectorMetric.DOT_PRODUCT,
-        service={"provider": "nvidia", "modelName": "NV-Embed-QA"},
-    )
-    yield collection
-
-    sync_database.drop_collection(TEST_SERVICE_COLLECTION_NAME)
-
-
 @pytest.fixture(scope="function")
 def sync_empty_collection(sync_collection: Collection) -> Iterable[Collection]:
     """Emptied for each test function"""
     sync_collection.delete_all()
     yield sync_collection
-
-
-@pytest.fixture(scope="function")
-def sync_empty_service_collection(
-    sync_service_collection: Collection,
-) -> Iterable[Collection]:
-    """Emptied for each test function"""
-    sync_service_collection.delete_all()
-    yield sync_service_collection
 
 
 @pytest.fixture(scope="function")
@@ -135,14 +100,6 @@ def async_empty_collection(
 ) -> Iterable[AsyncCollection]:
     """Emptied for each test function"""
     yield sync_empty_collection.to_async()
-
-
-@pytest.fixture(scope="function")
-def async_empty_service_collection(
-    sync_empty_service_collection: Collection,
-) -> Iterable[AsyncCollection]:
-    """Emptied for each test function"""
-    yield sync_empty_service_collection.to_async()
 
 
 __all__ = [
