+++ tests/idiomatic/integration/test_dml_async.py
@@ -34,8 +34,6 @@ from astrapy.operations import (
 )
 from astrapy.ids import ObjectId, UUID
 
-from ..conftest import is_vector_service_available
-
 
 class TestDMLAsync:
     @pytest.mark.describe("test of collection count_documents, async")
@@ -1314,103 +1312,6 @@ class TestDMLAsync:
         assert set(resp_pr2.keys()) == {"f"}
         await acol.delete_all()
 
-    @pytest.mark.skipif(
-        not is_vector_service_available(), reason="No 'service' on this database"
-    )
-    @pytest.mark.describe("test of vectorize in collection methods, async")
-    async def test_collection_methods_vectorize_async(
-        self,
-        async_empty_service_collection: AsyncCollection,
-    ) -> None:
-        acol = async_empty_service_collection
-
-        await acol.insert_one({"t": "tower"}, vectorize="How high is this tower?")
-        await acol.insert_one({"t": "vectorless"})
-        await acol.insert_one({"t": "vectorful"}, vector=[0.01] * 1024)
-
-        await acol.insert_many(
-            [{"t": "guide"}, {"t": "seeds"}],
-            vectorize=[
-                "This is the instructions manual. Read it!",
-                "Other plants rely on wind to propagate their seeds.",
-            ],
-        )
-        await acol.insert_many(
-            [{"t": "dog"}, {"t": "cat_novector"}, {"t": "spider"}],
-            vectorize=[
-                None,
-                None,
-                "The eye pattern is a primary criterion to the family.",
-            ],
-            vectors=[
-                [0.01] * 1024,
-                None,
-                None,
-            ],
-        )
-
-        doc = await acol.find_one(
-            {},
-            vectorize="This building is five storeys tall.",
-            projection={"$vector": False},
-        )
-        assert doc is not None
-        assert doc["t"] == "tower"
-
-        docs = [
-            doc
-            async for doc in acol.find(
-                {},
-                vectorize="This building is five storeys tall.",
-                limit=2,
-                projection={"$vector": False},
-            )
-        ]
-        assert docs[0]["t"] == "tower"
-
-        rdoc = await acol.find_one_and_replace(
-            {},
-            {"t": "spider", "$vectorize": "Check out the eyes!"},
-            vectorize="The disposition of the eyes tells much",
-            projection={"$vector": False},
-        )
-        assert rdoc["t"] == "spider"
-
-        r1res = await acol.replace_one(
-            {},
-            {"t": "spider", "$vectorize": "Look at how the eyes are placed"},
-            vectorize="The disposition of the eyes tells much",
-        )
-        assert r1res.update_info["nModified"] == 1
-
-        udoc = await acol.find_one_and_update(
-            {},
-            {"$set": {"$vectorize": "Consider consulting the how-to"}},
-            vectorize="Have a look at the user guide...",
-            projection={"$vector": False},
-        )
-        assert udoc["t"] == "guide"
-
-        u1res = await acol.update_one(
-            {},
-            {"$set": {"$vectorize": "Know how to operate it before doing so."}},
-            vectorize="Have a look at the user guide...",
-        )
-        assert u1res.update_info["nModified"] == 1
-
-        ddoc = await acol.find_one_and_delete(
-            {},
-            vectorize="Some trees have seeds that are dispersed in the air!",
-            projection={"$vector": False},
-        )
-        assert ddoc["t"] == "seeds"
-
-        d1res = await acol.delete_one(
-            {},
-            vectorize="yet another giant construction in this suburb.",
-        )
-        assert d1res.deleted_count == 1
-
     @pytest.mark.describe("test of ordered bulk_write, async")
     async def test_collection_ordered_bulk_write_async(
         self,
@@ -1494,42 +1395,6 @@ class TestDMLAsync:
             AsyncUpdateOne({}, {"$set": {"b": 1}}, vector=[1, 15]),
             AsyncReplaceOne({}, {"a": 10}, vector=[5, 6]),
             AsyncDeleteOne({}, vector=[-8, 7]),
-        ]
-        await acol.bulk_write(bw_ops)
-        found = [
-            {k: v for k, v in doc.items() if k != "_id"}
-            async for doc in acol.find({}, projection=["a", "b"])
-        ]
-        assert len(found) == 2
-        assert {"a": 10} in found
-        assert {"a": 2, "b": 1} in found
-
-    @pytest.mark.skipif(
-        not is_vector_service_available(), reason="No 'service' on this database"
-    )
-    @pytest.mark.describe("test of bulk_write with vectorize, async")
-    async def test_collection_bulk_write_vectorize_async(
-        self,
-        async_empty_service_collection: AsyncCollection,
-    ) -> None:
-        acol = async_empty_service_collection
-
-        bw_ops = [
-            AsyncInsertOne({"a": 1}, vectorize="The cat is on the table."),
-            AsyncInsertMany(
-                [{"a": 2}, {"z": 0}],
-                vectorize=[
-                    "That is a fine spaghetti dish!",
-                    "I am not debating the effectiveness of such approach...",
-                ],
-            ),
-            AsyncUpdateOne(
-                {},
-                {"$set": {"b": 1}},
-                vectorize="Oh, I love a nice bolognese pasta meal!",
-            ),
-            AsyncReplaceOne({}, {"a": 10}, vectorize="The kitty sits on the desk."),
-            AsyncDeleteOne({}, vectorize="I don't argue with the proposed plan..."),
         ]
         await acol.bulk_write(bw_ops)
         found = [
