+++ tests/idiomatic/integration/test_exceptions_sync.py
@@ -28,7 +28,7 @@ from astrapy.exceptions import (
     TooManyDocumentsToCountException,
 )
 
-from ..conftest import AstraDBCredentials, is_vector_service_available
+from ..conftest import AstraDBCredentials
 
 
 class TestExceptionsSync:
@@ -273,23 +273,6 @@ class TestExceptionsSync:
             )
 
         sync_database.drop_collection(TEST_LOCAL_COLLECTION_NAME)
-
-    @pytest.mark.skipif(
-        not is_vector_service_available(), reason="No 'service' on this database"
-    )
-    @pytest.mark.describe(
-        "test of database create_collection dimension-mismatch failure, sync"
-    )
-    def test_database_create_collection_dimension_mismatch_failure_sync(
-        self,
-        sync_database: Database,
-    ) -> None:
-        with pytest.raises(DataAPIResponseException):
-            sync_database.create_collection(
-                "collection_name",
-                dimension=123,
-                service={"provider": "nvidia", "modelName": "NV-Embed-QA"},
-            )
 
     @pytest.mark.describe("test of database one-request method failures, sync")
     def test_database_method_failures_sync(
